Updated splitting algorithm
The new splitting algorithm goes over the items, determines a best
estimate for the duration and then adds it to the smallest group.

The are edge-cases that this algorithm can't handle properly. For
example, when the last test item has a large test duration, the
current solution won't create the optimal solution. This can be
improved by sorting the test durations first, and starting with assignment
of tests that have the largest test duration.